feat(data_storage): store google extracted data to database.
All the collected data can now be stored to the database, this also comes with ability to
render the html into java object and table entities that can then be used later for data processing.

If a user was searched before, we nolonger query from google but from the database itself as we now the results locally.

Issues:

Still need to improve the way I handle exceptions and project structure.